+++ kubernetes-client/src/main/java/io/fabric8/kubernetes/client/DefaultKubernetesClient.java
@@ -342,6 +342,11 @@ public class DefaultKubernetesClient extends BaseClient implements NamespacedKub
   }
 
   @Override
+  public AdmissionRegistrationAPIGroupDSL admissionRegistration() {
+    return adapt(AdmissionRegistrationAPIGroupClient.class);
+  }
+
+  @Override
   public AppsAPIGroupDSL apps() {
     return adapt(AppsAPIGroupClient.class);
   }